Create ChangePasswordModal component
